+++ config/routes.rb
@@ -201,8 +201,6 @@ Rails.application.routes.draw do
     end
   end
 
-  resources :hosting_requests, path: "hosting-requests"
-
   resources :images do
     collection do
       post :upload, action: :create_multiple
@@ -314,13 +312,6 @@ Rails.application.routes.draw do
     # post "contribute/tool/submit", action: "contribute_tool_submit", as: :contribute_tool_submit
 
     get "tool/request", action: "tool_request", as: :tool_request
-
-    get "dataset/hosting", to: redirect("dataset/hosting/start")
-    get "dataset/hosting/start", action: "dataset_hosting_start", as: :dataset_hosting_start
-    post "dataset/hosting/start", action: "dataset_hosting_set_description", as: :dataset_hosting_set_description
-    post "dataset/hosting", action: "dataset_hosting_register_user", as: :dataset_hosting_register_user
-    patch "dataset/hosting", action: "dataset_hosting_sign_in_user", as: :dataset_hosting_sign_in_user
-    get "dataset/hosting/submitted", action: "dataset_hosting_submitted", as: :dataset_hosting_submitted
   end
 
   scope module: :search do