+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 ### Refactoring
 - Removed deprecated challenges
 - Merged community tools and tools codebase
+- Removed deprecated hosting requests
 
 ### Bug Fix
 - Fixed an issue displaying variable subfolders for a folder with parentheses